Make into cookie cutter tempalte
