--- .bsp/sbt.json
@@ -1,1 +1,0 @@
-{"name":"sbt","version":"1.5.5","bspVersion":"2.0.0-M5","languages":["scala"],"argv":["/Users/kelvinmacartney/.sdkman/candidates/java/14.0.2.j9-adpt/bin/java","-Xms100m","-Xmx100m","-classpath","/Users/kelvinmacartney/Library/Application Support/JetBrains/Toolbox/apps/IDEA-U/ch-0/212.5284.40/IntelliJ IDEA.app.plugins/Scala/launcher/sbt-launch.jar","xsbt.boot.Boot","-bsp","--sbt-launch-jar=/Users/kelvinmacartney/Library/Application%20Support/JetBrains/Toolbox/apps/IDEA-U/ch-0/212.5284.40/IntelliJ%20IDEA.app.plugins/Scala/launcher/sbt-launch.jar"]}
