--- app/src/main/resources/application.conf
@@ -1,7 +1,0 @@
-ops-server-config {
-    port = 5002
-}
-
-rest-config {
-    port = 5024
-}
