--- app/src/main/scala/org/example/com/config/ApplicationConfig.scala
@@ -1,6 +1,0 @@
-package org.example.com.config
-
-final case class ApplicationConfig(opsServerConfig: OpsServerConfig, restConfig: RestApiConfig)
-
-final case class RestApiConfig(port: Int)
-final case class OpsServerConfig(port: Int)
