+++ cookiecutter.json
@@ -1,0 +1,8 @@
+{
+  "project_name": "Scala rest api http4s",
+  "project_slug": "scala-rest-api-http4s",
+  "project_subdomain": "foxi",
+  "_copy_without_render": ["project/build.properties", "project/plugins.sbt"],
+  "project_short_description": "Scala rest api written in http4s",
+  "author": "Kevo Macartney"
+}
