--- domain/src/main/scala/org/example/com/config/RateLimiter.scala
@@ -1,5 +1,0 @@
-package org.example.com.config
-
-import scala.concurrent.duration.Duration
-
-case class RateLimiter(rate: Int, duration: Duration)
