--- domain/src/main/scala/org/example/com/repositories/Repository.scala
@@ -1,10 +1,0 @@
-package org.example.com.repositories
-
-import java.util.UUID
-import scala.language.higherKinds
-
-trait Repository[F[_]] {
-  def get(id: String): F[Option[RepositoryItem]]
-}
-
-final case class RepositoryItem(name: String, id: UUID)
