--- test-support/src/main/scala/org/example/com/e2e/support/http/MockHttp.scala
@@ -1,16 +1,0 @@
-package org.example.com.e2e.support.http
-
-import java.net.ServerSocket
-
-object MockHttp {
-
-  protected def getFreePort: Int = {
-    val socket = new ServerSocket(0)
-    socket.setReuseAddress(true)
-
-    val localPort = socket.getLocalPort
-    socket.close()
-
-    localPort
-  }
-}
