+++ {{ cookiecutter.project_slug }}/.gitignore
