+++ {{ cookiecutter.project_slug }}/.scalafmt.conf
