+++ {{ cookiecutter.project_slug }}/app/src/main/resources/application.conf
@@ -1,0 +1,7 @@
+ops-server-config {
+    port = 5002
+}
+
+rest-config {
+    port = 8080
+}
