+++ {{ cookiecutter.project_slug }}/app/src/main/resources/secrets.json
