+++ {{ cookiecutter.project_slug }}/domain/src/main/scala/org/{{ cookiecutter.project_subdomain }}/com/config/RateLimiter.scala
@@ -1,0 +1,5 @@
+package org.{{ cookiecutter.project_subdomain }}.com.config
+
+import scala.concurrent.duration.Duration
+
+case class RateLimiter(rate: Int, duration: Duration)
