+++ {{ cookiecutter.project_slug }}/project/build.properties
