+++ {{ cookiecutter.project_slug }}/project/plugins.sbt
